native timesincelastkeypress function used
Wasn't aware, but roku already has a time since last keypress function and it runs on seconds. w00t

So reduced the code down, and removed the keep alive timer. This is the way it should be. Now it is fully dependent on the last button pressed on the remote.